Eliminate user_page_limit global symbol.
The name "user_page_limit" did not follow the naming convention, which
stated that its name should start with palloc_ since it is defined in
palloc.c.  But it was only used in one function, so it was better off
as a function parameter anyhow.

Found by Godmar with his process-linker-map.pl.